+++ smarts/env/wrappers/metric/counts.py
@@ -1,6 +1,4 @@
-# MIT License
-#
-# Copyright (C) 2021. Huawei Technologies Co., Ltd. All rights reserved.
+# Copyright (C) 2022. Huawei Technologies Co., Ltd. All rights reserved.
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -19,29 +17,25 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-from collections import namedtuple
 
-from shapely.geometry import LineString
-from shapely.ops import split
-
-# TODO: Move these into SMARTS and reuse from tests
-BubbleGeometry = namedtuple(
-    "BubbleGeometry",
-    ["bubble", "airlock_entry", "airlock_exit", "airlock"],
-)
+from dataclasses import dataclass
 
 
-def bubble_geometry(bubble, road_map):
-    bubble_geometry_ = bubble.zone.to_geometry(road_map)
-    airlock_geometry = bubble_geometry_.buffer(bubble.margin)
-    split_x, split_y = airlock_geometry.centroid.coords[0]
-    divider = LineString([(split_x, -999), (split_x, split_y + 999)])
-    airlock_entry_geometry, airlock_exit_geometry = (
-        split(airlock_geometry, divider)
-    ).geoms
-    return BubbleGeometry(
-        bubble=bubble_geometry_,
-        airlock_entry=airlock_entry_geometry,
-        airlock_exit=airlock_exit_geometry,
-        airlock=airlock_geometry,
-    )
+@dataclass(frozen=True)
+class Counts:
+    """Performance count values."""
+
+    goals: int = 0
+    """ Number of episodes completed successfully by achieving the goal.
+    """
+    episodes: int = 0
+    """ Number of episodes traversed.
+    """
+    steps: int = 0
+    """ Sum of steps taken over all episodes.
+    """
+    steps_adjusted: int = 0
+    """ Sum of steps taken over all episodes. The number of steps, in an episode
+    where the agent did not achieve the goal, is replaced with the agent's  
+    max_episode_steps value.
+    """